telegram/pyproject.toml
[tool.poetry]
name = "telegram"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.14.0"
pandas = "^2.2.3"
python-dotenv = "^1.0.1"
openpyxl = "^3.1.5"
numpy = "^2.1.3"
loguru = "^0.7.2"
redis = "^5.2.0"
typing-extensions = "^4.12.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
